push wip kv refactor

+++ src/cute_base64.cpp
@@ -53,12 +53,12 @@ static const int s_base64_to_6bits[79] = {
 	26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50,
 };
 
-error_t base64_encode(void* dst, int dst_size, const void* src, int src_size)
+error_t base64_encode(void* dst, size_t dst_size, const void* src, size_t src_size)
 {
-	int out_size = CUTE_BASE64_ENCODED_SIZE(src_size);
+	size_t out_size = CUTE_BASE64_ENCODED_SIZE(src_size);
 	if (dst_size < out_size) return error_failure("`dst` buffer too small to place encoded output.");
 
-	int triplets = (src_size) / 3;
+	size_t triplets = (src_size) / 3;
 	int pads = (src_size) % 3 ? 3 - (src_size) % 3 : 0;
 
 	const uint8_t* in = (const uint8_t*)src;
@@ -121,11 +121,11 @@ error_t base64_encode(void* dst, int dst_size, const void* src, int src_size)
 	return error_success();
 }
 
-error_t base64_decode(void* dst, int dst_size, const void* src, int src_size)
+error_t base64_decode(void* dst, size_t dst_size, const void* src, size_t src_size)
 {
 	if (!src_size) return error_success();
 	if (src_size % 4) return error_failure("`src_size` is not a multiple of 4 (all base64 streams must be padded to a multiple of four with `=` characters).");
-	int quadruplets = src_size / 4;
+	size_t quadruplets = src_size / 4;
 	
 	const uint8_t* in = (const uint8_t*)src;
 	uint8_t* out = (uint8_t*)dst;
